# openapi-httpexchange-sdk-generator/openapi/httpexchange-server@0.0.2-SNAPSHOT.json
{"openapi":"3.1.0","info":{"title":"用户管理服务 API","description":"HttpExchange Server - 提供用户管理的核心接口服务","contact":{"name":"开发团队","email":"dev@example.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8181","description":"本地开发服务器"}],"tags":[{"name":"user-controller","description":"用户管理控制器\n \n <p>提供用户的增删改查 RESTful API 接口</p>"}],"paths":{"/api/users/{id}":{"get":{"tags":["user-controller"],"summary":"根据ID获取用户详情","description":"根据ID获取用户详情","operationId":"getUserById","parameters":[{"name":"id","in":"path","description":"用户唯一标识","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"用户详情，如果不存在返回404","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"put":{"tags":["user-controller"],"summary":"更新用户信息","description":"更新用户信息","operationId":"updateUser","parameters":[{"name":"id","in":"path","description":"用户唯一标识","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"description":"更新用户请求，包含需要更新的字段","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"200":{"description":"更新后的用户信息，如果用户不存在返回404","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}},"delete":{"tags":["user-controller"],"summary":"删除用户","description":"删除用户","operationId":"deleteUser","parameters":[{"name":"id","in":"path","description":"用户唯一标识","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"删除成功返回204，用户不存在返回404"}}}},"/api/users":{"get":{"tags":["user-controller"],"summary":"获取所有用户列表","description":"获取所有用户列表","operationId":"getAllUsers","responses":{"200":{"description":"用户列表","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"post":{"tags":["user-controller"],"summary":"创建新用户","description":"创建新用户","operationId":"createUser","requestBody":{"description":"创建用户请求，包含用户名、邮箱等信息","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"创建成功的用户信息","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}}},"components":{"schemas":{"UpdateUserRequest":{"type":"object","description":"更新用户请求\n \n <p>用于更新用户信息时提交的数据</p>","properties":{"username":{"type":"string","description":"用户名，可选，长度3-50个字符"},"email":{"type":"string","description":"邮箱地址，可选，需符合邮箱格式"},"fullName":{"type":"string","description":"用户真实姓名，可选"},"status":{"type":"string","description":"用户状态：ACTIVE-活跃，INACTIVE-未激活，DISABLED-已禁用"}}},"User":{"type":"object","description":"用户实体类\n\n <p>表示系统中的用户信息</p>","properties":{"id":{"type":"integer","format":"int64","description":"用户唯一标识"},"username":{"type":"string","description":"用户名，用于登录和显示"},"email":{"type":"string","description":"用户邮箱地址"},"fullName":{"type":"string","description":"用户真实姓名"},"status":{"type":"string","description":"用户状态：ACTIVE-活跃，INACTIVE-未激活，DISABLED-已禁用"},"createdAt":{"type":"string","format":"date-time","description":"用户创建时间"},"updatedAt":{"type":"string","format":"date-time","description":"用户信息最后更新时间"}}},"CreateUserRequest":{"type":"object","description":"创建用户请求\n \n <p>用于创建新用户时提交的数据</p>","properties":{"username":{"type":"string","description":"用户名，必填，长度3-50个字符"},"email":{"type":"string","description":"邮箱地址，必填，需符合邮箱格式"},"fullName":{"type":"string","description":"用户真实姓名，可选"}}}}}}